requirements.txt
numpy
Pillow

https://coreui.io/vue/docs/components/accordion.html
